+++ go.work
@@ -16,7 +16,6 @@ use (
 	./taco/cmd/statesman
 	./taco/cmd/taco
 	./taco/internal
-	./taco/models
 	./taco/pkg/sdk
 	./taco/providers/terraform/opentaco
