+++ taco/cmd/taco/go.mod
@@ -1,6 +1,6 @@
 module github.com/diggerhq/digger/opentaco/cmd/taco
 
-go 1.24.0
+go 1.24
 
 require (
 	github.com/diggerhq/digger/opentaco/internal v0.0.0-00010101000000-000000000000
