+++ taco/internal/repositories/unit_repository.go
@@ -7,6 +7,7 @@ import (
 	"log"
 	"time"
 
+	"github.com/diggerhq/digger/opentaco/internal/domain"
 	"github.com/diggerhq/digger/opentaco/internal/query/types"
 	"github.com/diggerhq/digger/opentaco/internal/storage"
 	"gorm.io/gorm"
