--- taco/models/go.mod
@@ -1,14 +1,0 @@
-module github.com/diggerhq/digger/opentaco/models
-
-go 1.24.0
-
-require (
-	github.com/google/uuid v1.6.0
-	gorm.io/gorm v1.31.0
-)
-
-require (
-	github.com/jinzhu/inflection v1.0.0 // indirect
-	github.com/jinzhu/now v1.1.5 // indirect
-	golang.org/x/text v0.20.0 // indirect
-)
